requirements.txt
anyio==3.7.1
boltons==23.0.0
boto3==1.28.7
botocore==1.31.7
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==41.0.2
einops==0.6.0
environs==9.5.0
exceptiongroup==1.1.2
fastapi==0.89.1
filelock==3.12.2
fsspec==2023.6.0
gmpy2==2.1.2
h11==0.14.0
huggingface-hub==0.16.4
humanfriendly==10.0
humanize==4.7.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
mpmath==1.3.0
multiprocessing-logging==0.3.4
networkx==3.1
ninja==1.11.1
numpy==1.25.1
packaging==23.1
pip==23.0.1
pluggy==1.2.0
plumbum==1.8.2
protobuf==4.22.3
pycparser==2.21
pydantic==1.10.11
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.1.1b2
regex==2023.6.3
requests==2.31.0
rpyc==5.3.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
sentencepiece==0.1.99
setuptools==65.6.3
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
sympy==1.12
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.0
tqdm==4.65.0
transformers==4.29.2
triton==2.1.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.19.0
uvloop==0.17.0
wheel==0.40.0
zstandard==0.19.0
safetensors==0.3.1
Pillow==10.1.0
